# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Receptionist API",
        "description": "AI-powered voice receptionist system for hospital appointment management. This API handles doctor schedules, patient information, appointment booking, and slot management. Some endpoints require API key authentication via the X-API-Key header (appointments and patient lookup).",
        "contact": {
            "name": "API Support",
            "email": "support@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Alternative Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get doctor details",
                "description": "Retrieve detailed information about a specific doctor",
                "operationId": "37105695af3266d77b88e448686570c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "dr.smith@hospital.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+1234567890"
                                                },
                                                "specialization": {
                                                    "type": "string",
                                                    "example": "General Practitioner"
                                                },
                                                "department": {
                                                    "type": "string",
                                                    "example": "General Medicine"
                                                },
                                                "slots_per_appointment": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "max_appointments_per_day": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/doctors/{id}/availability": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Check doctor availability",
                "description": "Check if a doctor is available on a specific date",
                "operationId": "c376af00b4a530d40e269df870f1cd4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-09"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Availability checked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "available": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "available": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Doctor is not available on this date"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error message"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key authentication. Use one of the configured API keys in the X-API-Key header.",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}